+++ bin/check_unused_dependencies.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+cargo +nightly udeps
+cargo +nightly udeps --all-targets